retry binding port
When restarting or reloading a server, there's a bit of a race between the old
server shutting down its listening socket and the new server binding the port.

This makes the server retry binding the port for 10 seconds before giving up,
which is hopefully always more time than the old one needs to shut down.

Swift does something similar.