cache Copy() temp buffers
This change reduces GC pauses by about 75% under simulated load.

It includes a cache pool for object re-use, and a replacement for io.Copy
that reuses its temporary buffers.  This Copy also takes multiple writers.

I've been benchmarking cache pool implementations for a few days, and this
is the fastest thing I've come up with.  But the interface is pretty generic,
so it can easily be improved later.  I know gholt et al are working on this
problem as well.