work on xattr loading
When we already have an open file (GETs), loading metadata with fgetxattr and
the file descriptor is ~twice as fast as loading it using getxattr and the
filename.

    BenchmarkFilename    18933 ns/op
    BenchmarkFD           9169 ns/op

Also, I sped up the loading in the original case by changing how space for the
attributes is allocated.  This was the benchmark before I started.

    BenchmarkFilename    40091 ns/op

So this only shaves ~.03 ms off GETs on my VM, but optimizations are hard to
find.