limit replicate calls
Limit "REPLICATE" calls to a concurrency of 5 per disk.
Limit all other calls to 25 per disk.

Refactor limiting code to be usable for both.
Get rid of atomic stuff, because it wasn't ever really right anyway.

I suspect this is going to get more complicated in the future, so I'm trying
not to over-design it for now.  For example, I'd like to concurrency limit
entire replication passes as a unit, not just REPLICATE calls.  I also think
we might want to limit PUT/DELETE and GET/HEAD separately.