super non-fancy router
I implemented a dumb router I knew would work, then optimized it a lot.
Hopefully we can swap it for a third party router later.

It's about the same speed as httptreemux for the number of routes we have -
actual routing time is negligible for both, and they're kind of dominated
by the time it takes to allocate and register the "vars" map.

    BenchmarkRouteObject        1000000   2310 ns/op   352 B/op   3 allocs/op
    BenchmarkRouteObjectTree    1000000   2378 ns/op   352 B/op   3 allocs/op