go: fix unpickling problem
There's a problem where the pickle stack winds up containing a subslice of
itself when creating a new list type object.  Since they point to the same
memory, when it modifies the stack, it also modifies the item in the stack,
which is wrong.

Fix it by making a copy of that subslice and pushing it instead.

I looked around, but nothing else seems to retain a reference to state.mark().

Change-Id: Ia94b1f73d3e50e6597f2e77da5bf0306a4a6a8c1