go: fix expecttransport memory usage
DumpRequestOut wasn't really working.  It's meant to be a debugging function,
and it just writes the whole request, body and all, into a buffer, then strips
out the headers.  That means it can use a ton of memory.  So this just
serializes the request ourselves, which is bad and ugly, but at least it kind
of works.

Change-Id: Iafd4ec8bf7ffbf2506310722853b055d422228f1