go: pass vars to container update
Since we kick container updates out to a background goroutine and let the
request handler return, the vars context can be cleared before GetVars is
called.

That means means it doesn't know what device the request was for when saving
asyncs, and it just dumps it into the device root.

Change-Id: I8fff740fa470b9d5012864ed92c8d0767efcb519