go: proxy rewrite
I know this is a huge change.  The proxy still isn't very advanced,
this just cleans it up and lays some framework down.

Most functionality was moved out to a direct client.
There's a proxy context to hold auth info and per-request caches.
Better tempauth.  I mean it's still not great, but it's better.
Inklings of callback-based authentication.
Account and container info fetching and caching.

Change-Id: I2543040a5f9faed20e713eb24e2189801232b76e