go: have replicateDevice go routine loop per device
This will prevent a slow disk from slowing down replication
on all the rest of the server.

This also changes the meaning of the replicator recon report. Since
we won't be waiting for the replicator to complete the whole
server the object_replication_time will be the average of
the replication time of all the devices. It won't be set until
there has been a complete pass on all the devices. This is a
little strange but I'm leaving it there for some backwards
compatibility.

The previous whole device replication times were a little misleading if
most the drives are done and it is still running on a few slow drives.
The new model makes the replication stats per drive and instead of
reporting them in recon it has a new API call to the replicator's
http server at /progress .  I could also add these per-drive stats
into recon but i will leave that as a further patch if somebody
wants it. I'm going to try to use the /progress call to get what I need.

The object replicator itself will monitor the progress of the
replicators and restart the go routines if one of them has not
reported for 4 hours.

Change-Id: Ifa5defbe684329979a533b558a57c38ac029ada1