Fix the removal of temporary files of tests
I noticed that /tmp had a number of leftover temporary files
after a run of  go test ./... . They resulted from 3 root causes
with separate fixes:

- Add a few defer before ts.Close() after makeObjectServer()
- Remove correct files in TestGetFileBadMetadata(), was copypasta
- Wait for goroutine in SwiftObject.Commit() to finish

The last one is not easy to fix, because it requires waiting for
Commit() in Close(), and that has performance implications for
real work. But we can fix the first two at least and possibly
revisit the wild goroutine problem later.

Change-Id: I30eea7fc9b8ff1e2857d201b07007e941626c546