go: fix replicator policy bugs / big refactor
The replicator was pretty broken with multiple poliices.

I thought the best fix was to restructure the replicator to have a
ReplicationDevice that holds all information per (device, policy).

This version is harder to mess up, more testable, and points at the
possibility of implementing a different ReplicationDevice for other
policies later if they want to plug into the replicator.

Change-Id: If0a24a5cdf3b9423cb37cd521cf22960e6f38e23