go: contain X-Trans-Id in response headers
Current the hummingbird proxy does not send back the X-Trans-Id while
the Python proxy does. There are two ways to set it in the response.

1. Set it in every HTTP method handler
2. Set in it a middleware

The latter one should be preferred since it requires minimal code
implementation. Since the transaction id is generated in the LogRequest
middleware, it is reasonable to set the X-Trans-Id header there.

Change-Id: I29b8c315402b8d612c6cd7b410c850628c8d7b72