pickle support for builder files
If a tuple is used as a dict key, promote it to a PickleTuple struct
so it can act as a map key in go.

Support loading "array.array" objects to PickleArrays.  The way this
works is kind of messy because pickles are gross.

Also, fix for BINFLOAT endianness, since it was wrong.  Why are
integers stored little-endian and floats big-endian?  ¯\_(ツ)_/¯

fixes #8