make current swift functional tests pass
Fix some new tests/functionality from swift from the last few months.

If-Match and If-None-Match needed to be a little more robust, and to
know about the X-Backend-Etag-Is-At header.

A Content-Range header needed to be added to 416 responses.

Container GET/HEAD needed a Last-Modified header, and that needs to
be updated on container POST.