go: more region-smart replication
When syncing a file to multiple devices in a remote region, only
actually sync with one of them. For handoff replication, do an
in-sync check on the others so we can still remove objects that are
fully synced.  For non-handoff data, just send a ping to keep the
connection alive.

this is a port of:

https://review.openstack.org/#/c/339296/