Add structured logging using zap.
Switching from syslog based logging to structured logging.
Currently all the logs will logged in specified log_path.
In future we plan to use upstart/systemd to redirect logs and daemonize.

Also added /loglevel endpoint to all the server that can report on or change the current
logging level. GET requests return a JSON description of the current logging level. PUT
requests change the logging level and expect a payload like:
  {"level":"info"}
It's safe to change the logging level while the server is running.