Added policy code to proxy server's direct client
ProxyDirectClient lost its ability to directly work with objects, instead it
delegates that to proxyObjectClient implementations. There's just one
implementation right now, but imagine more later based on policies.

In order to know which proxyObjectClient implementation to pick, the
ProxyDirectClient needs to know the policy. There was already code in the
middleware section for retrieving and caching ContainerInfo, so I wanted to
leverage that code and cache. Cyclic imports, etc. meant a moderate amount of
refactoring to get this done.

So now there are a couple of interfaces, ContainerInfo and ContainerInfoSource.
The prior ContainerInfo struct needed to be unexported and have methods
attached to it instead. Pretty straighforward conversion there.

When the proxy server creates a ProxyDirectClient now, it passes in a
ContainerInfoSource which the client uses to figure out the policy and
proxyObjectClient implementation.

Finally, the ProxyServer was storing the ProxyDirectClient before, but I just
moved that into the ProxyContext now, as created by the ProxyContextMiddleware.
This was needed as the context isn't known until then in order to pass it to
the ProxyDirectClient.

Oh, and I added GrepObject to the ProxyClient interface just because it was
implemented before but not in the interface.