fix xlo range requests
I guess that change was supposed to fix some smoke test.  But it broke
a bunch of functional tests.  I think it was trying to output the whole
object when the range header wasn't empty, but didn't have any parseable
ranges.  So I just did that while fixing the functional tests.