proxy client changes
The proxy client never re-uses container info.  There's a place to
cache it, and it's checked, but it's never set.  This means that the
container info can be fetched from memcache 4 or 5 times per request.

I also took the opportunity to rework things so we don't create
new proxy object clients for every operation.  They are created once,
when the proxy client is created.