HUM-261: proxy side of EC PUT
Allow the EC scheme be configured in the policy config.

Limit the number of nodes that are written by the proxy using the
object client, which is configured by the policy.

In a world where EC is just another policy, this seems less weird to
me than building the EC scheme into the ring.