HUM-390: ec ephemeral port usage and nil crash
Fix a couple of show stopper bugs in the initial EC code.

Ephemeral ports are being exhausted by runaway time-waits, hopefully
mitigated by increasing the max idle conns setting (like we've had to
do everywhere).  Also fixing one spot where I failed to discard a
response body, so the connection couldn't be reused.

A nil crash cauesd by ecSplit modifying the slice it was given as an
argument and the caller not expecting that.