HUM-415 Better systemd support
Type=forking did not pan out. systemd is quite picky. If you use
Type=forking and PIDFile=x it does not expect a reload to cause a PID
change. I tried a few ways "suggested on the internet" such as ensuring
the PIDFile was fully updated before exiting whatever ExecReload called
to no avail. Also, I had no idea how to make a graceful stop work under
Type=forking, which made systemctl restarts pointless. So I gave up,
went back to to thinking board, put on my drawing cap, and...

...now we have this. Type=simple still, but systemd is now calling a new
systemd controller process of Hummingbird, which then translates what
systemd wants to happen with what we want to happen.

start: starts hummingbird up with the args given, then listens for
signals or process exit. If it gets a signal, it passes that on to the
real hummingbird process. If that signal was HUP, it then loops back
around relaunches the real hummingbird process (therefore leaving the
original PID running, which makes systemd happy).

reload: sends the HUP signal to our systemd helper process.

stop: sends the TERM signal to our systemd helper process, and waits up
to five seconds for that process to exit. systemd expects the main
process to have exited when stop returns or it will kill it. When our
helper receives TERM it ill forward the signal to the real hummingbird
process and immediately exit. This has the downside that a weird hung
real hummingbird process won't be detected by systemd; we may need to
improve this in the future to launch a "make sure it dies" process, but,
I guess wait until it becomes an issue.

I tested this in various ways I could think of, the main interesting one
being a slow upload that completely survived a reload of all processes.
If you want to do this yourself, I can recommend this for the slow
upload:

for i in {{1..100}} ; do sleep 1 ; echo . ; done | nectar put c/o

When all the services are reloaded while that upload is in progress, you
should see (on an haio) double proxy processes as well as 3 extra object
server processes. Once the upload completes, you should see these extra
processes go away and "graceful exit" log messages come through.