HUM-518 Andrewd Quarantine Repairman
This adds a background job to andrewd that will scan the whole cluster
for quarantined items and repair them. For items that it has the name
for, it can repair just that item by pulling from a node that has a
copy. For items that it just knows the hash of, it can queue up a
replication pass for that partition.

There are some ancillary changes in here too:

The recon quarantine detail responses had to start including the device
data as well.

I needed to add PartitionForHash to the ring since sometimes the only
thing we know about a quarantined item is what its ring hash was, so the
best we can do is ask for a full replication cycle for the corresponding
partition.

I added a getDB to AutoAdmin. I'd like to move all the DB instantiation
stuff there, such as out of drivewatch, but I'll leave that for the
future. I also added an HTTP client to AutoAdmin and I'd like to get
everything in andrewd using the same client.

I added a new replication_queue table to the andrewd database, but
nothing actually acts on the entries there, yet.