Updated from review
repairItem is now repairObject since that's what it does.

Repairs of accounts and containers are skipped as they don't have
priority replication at this time (they don't need it since replication
passes are pretty fast). Left the rest of the code in there though; it
will keep the quarantined directories clean at the least.

Also, added a couple more debug lines indicating when a pass starts and
completes.