HUM-591 Better Dispersion Report
Dropped the service_error and device_error tables. My idea of simple
tables to encompass all error possibilities was a bit... fantastic.

Instead, there's a dispersion_scan_failure for the errors we can have
_right now_ and we'll make more tables for other error classes later.

So this dropped a lot of stuff for tracking those generic errors and
instead added code for tracking these dispersion scan failures, which
are reported in detail with the recon -d report.

I also took the opportunity to format things a bit better and to print
out the first few partitions that only have one or zero copies left.

We'll have to modify that last "one or zero" part for EC once we get
there to instead be "I can't just barely construct the object" and "I
can't construct the object anymore".