HUM-611 : EC posts
added new column to object table- you need to reset cluster after this patch

a post to a stabilized object will set the row's restabilized=1 which will
be looked at be the nursery stabilizer. the stabilzer will send ec-shard POSTs
to all the copies of the shard to verify it is completely distributed and
if so will set that flag back to 0.