HUM-451 Dispersion scan support for HEC
When there's a HEC policy, the shards of the dispersion objects will now
be scanned for and replication jobs queued for missing shards.

Before, it would just query each of the many devices from ring.GetNodes
for the full object, and every single one would respond successfully if
there had been at least data:shards in place. Lots of extra work and
useless in the end.

One side effect on this new change is that before the dispersion objects
get stabilized it will seem to the scanner that all the shards are
missing (since they essentially are) and it will queue up a bunch of
wasted effort replication jobs. In practice this shouldn't be a big deal
as it will only happen right after a dispersion populate and before
stabilization, and the cluster or policy should be brand new with near
zero data so the wasted effort will be low.