HUM-676 Global Clusters
Which is a pretty grandiose name for using slightly different node
sorting strategies.  Implemented (mostly) as described here:

https://docs.openstack.org/swift/ocata/overview_global_cluster.html

This lets you set read and write affinity.  Nobody should probably
ever use write affinity, but it is included for feature parity and
because it's not that complicated.

I didn't implement sorting_method, it just uses read affinity if it's
set.