HUM-722 Andrewd unmounted monitor optimization
Now when a device is cleared from a ring, or attempted to be cleared
from a ring, any more attempts to clear it from that ring will be
ignored for a while (4 hours by default).

This is needed because the servers report all unmounted devices for
forever and forever trying to remove them from the builders is more
wasteful than I thought. Loading those builders up just to check if the
device is in the ring takes a lot more time than I thought, especially
with 2**22 rings.

Anyway, we can't permanently ignore them after ensuring they're cleared
because someone could add them back in outside of Andrewd. So... ignore
durations!

Also, snuck in some dividebyzero fixes.