HUM-820 Add opentracing to memcache calls
Most of this change is propagating the context down to all the memcache
calls. The actual trace just creates one single span for the memcache client
call, adding the key as a tag if it's an account or container get. Auth keys
are not added as tags. Potential future additions/fixes are tagging errors
or times when we have to switch to a different server in the ring.