HUM-891 Fix context canceled error msg in proxy.
As we are using goroutine in quorumResponse & firstResponse and bailing out earlier,
passing the proxy context as it is down the stack causes "context cancelled" error.
This patch extract the tracing span from the proxy request's Context and
inject it into a new Context which is passed down the stack.