HUM-897 close some race conditions in account
Basically, make sure it grabs a flock on the hash dir *any* time it
deals with the database as a file on the filesystem.

Also, make sure that when it renames database files, any journals
move with database and old journals get removed.

When pulling a cached connection out of the pool, make sure the
underlying inode hasn't changed.